refactor(build): refactor build APIs and services
