+++ rolt/builds/apis/build_apis.py
@@ -6,17 +6,17 @@ from rest_framework.views import APIView
 
 from rolt.accounts.selectors.customer_selector import CustomerSelector
 from rolt.builds.models import Build
-from rolt.builds.selectors import build_check_duplicate_combo
-from rolt.builds.selectors import build_exists
-from rolt.builds.selectors import build_get_by_id
-from rolt.builds.selectors import customer_build_get_by_id
-from rolt.builds.selectors import customer_build_list
-from rolt.builds.selectors import preset_build_get_by_id
-from rolt.builds.selectors import preset_builds_list
-from rolt.builds.selectors import service_list_by_codes
-from rolt.builds.services import build_create
-from rolt.builds.services import build_delete
-from rolt.builds.services import build_update
+from rolt.builds.selectors.build_selectors import build_check_duplicate_combo
+from rolt.builds.selectors.build_selectors import build_exists
+from rolt.builds.selectors.build_selectors import build_get_by_id
+from rolt.builds.selectors.build_selectors import customer_build_get_by_id
+from rolt.builds.selectors.build_selectors import customer_build_list
+from rolt.builds.selectors.build_selectors import preset_build_get_by_id
+from rolt.builds.selectors.build_selectors import preset_builds_list
+from rolt.builds.selectors.service_selectors import service_list_by_codes
+from rolt.builds.services.build_services import build_create
+from rolt.builds.services.build_services import build_delete
+from rolt.builds.services.build_services import build_update
 from rolt.common.utils import inline_serializer
 from rolt.common.utils import user_in_group
 from rolt.components.selectors.keycap_selectors import keycap_get
